requirements.txt
# Core dependencies for IIT Rank Extractor
requests>=2.31.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.5
pandas>=2.0.0
python-dotenv>=1.0.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
numpy>=1.24.0

# Data processing and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Web scraping and API requests
beautifulsoup4>=4.11.0

# Database support
sqlalchemy>=1.4.0

# Logging and utilities
loguru>=0.6.0
tqdm>=4.64.0

# Development and notebook support
jupyter>=1.0.0